@@ -1,29 +1,28 @@
 module github.com/codeready-toolchain/toolchain-common
 
-go 1.18
+go 1.19
 
 require (
-	github.com/codeready-toolchain/api v0.0.0-20230308235149-3d2d5e3f0c16
-	github.com/emicklei/go-restful v2.16.0+incompatible // indirect
-	github.com/go-logr/logr v1.2.0
+	github.com/codeready-toolchain/api v0.0.0-20230417183849-f62ccbf3bd40
+	github.com/go-logr/logr v1.2.3
 	github.com/gofrs/uuid v3.3.0+incompatible
 	github.com/golang-jwt/jwt v3.2.2+incompatible
 	github.com/lestrrat-go/jwx v0.9.2
 	github.com/magiconair/properties v1.8.5
-	// using latest commit from 'github.com/openshift/api branch release-4.11'
-	github.com/openshift/api v0.0.0-20220912161038-458ad9ca9ca5
-	// using latest commit from 'github.com/openshift/library-go branch release-4.10'
-	github.com/openshift/library-go v0.0.0-20220630204433-c71d40c7de49
+	// using latest commit from 'github.com/openshift/api branch release-4.12'
+	github.com/openshift/api v0.0.0-20230213134911-7ba313770556
+	// using latest commit from 'github.com/openshift/library-go branch release-4.12'
+	github.com/openshift/library-go v0.0.0-20230301092340-c13b89190a26
 	github.com/pkg/errors v0.9.1
 	github.com/redhat-cop/operator-utils v1.3.3-0.20220121120056-862ef22b8cdf
 	github.com/stretchr/testify v1.7.0
 	gopkg.in/h2non/gock.v1 v1.0.14
 	gopkg.in/square/go-jose.v2 v2.3.0
 	gotest.tools v2.2.0+incompatible
-	k8s.io/api v0.24.2
-	k8s.io/apimachinery v0.24.2
-	k8s.io/client-go v0.24.2
-	sigs.k8s.io/controller-runtime v0.12.2
+	k8s.io/api v0.25.0
+	k8s.io/apimachinery v0.25.0
+	k8s.io/client-go v0.25.0
+	sigs.k8s.io/controller-runtime v0.13.0
 )
 
 require (
